Default settings
A new view-mode was added, but no settings about the field visibility.
This commit is just for making sure the feature doen't show as overridden in a brand new install, no considerations on whether the fields should be visible or not.

Field 'locked' settings are now also exported (because of features update http://drupal.org/node/1567506).