Stricter permissions for AJAX callbacks
Ajax callbacks currently require the `access content` permission. That permissions does not seem to be strict enough compared to who is supposed to use them and what the callbacks actually do.

I suggest that the permission is changed to `administer ddb_cp` which is also used for the form where the callbacks are intended to be used.