Reintroduce setting window size
Tests started failing because PhantomJS would render the page at mobile
resolutions, which hides some elements.

However the previous sizes failed in Selenium on Circle as it's
framebuffer doesn't seem to be that big.

So we compromise on a small desktop size that works on Circle, but
doesn't trigger mobile media types.