2304: Update Scheduler to 1.5
This fixes a security vulnerability in 1.2. 

Updating to 1.3 causes notices to be raised which we do not want. 
The patch for fixing this does not apply to 1.3. Consequently we
update to the latest version instead, 1.5. This also means we can
drop a patch which has been merged.

The new version introduces two new permissions which are updated
in secure permissions:

1. “schedule (un)publishing of nodes” is changed to “schedule 
publishing of nodes”
2. “view scheduled content” is introduces. This is added to roles
which already have “administer nodes” and “schedule 
publishing of nodes” permissions.

The new version may introduce new strings.

The new version introduces a few additional configuration options.
We use the defaults here.