2311: Update Entity API to 1.8
Release 1.5 contains a security vulnerability.

We want to get the most recent version as Entity API is a module 
which is used by other modules e.g. workflow and we do not want to
restrain out possibilities for using and updating such modules.

There are no changes to configuration, permissions or texts between 
the two versions.