2322: Update Panels to 3.9 and Ctools to 1.11
This fixes a security vulnerability in Panels <3.6.

Release notes for Panels recommends not updating to 3.6 due to 
regressions. Regressions are mentioned in release notes up until 3.9
which is the latest release. Consequently we update to that to avoid
introducing new problems.

This release requires updating to Ctools 1.11 so we do that as well.

This release adds three new permissions. These are added to the 
administrator role as suggested by the maintainer.

The module updates leads to configuration changes to all panels
exports.