2330: Update Media to 2.0 and File Entity to 2.0-beta3
This fixes multiple security vulnerabilities in the Media module.

File Entity needs to be updated as well as thumbnails do not work
with the combination of the new Media module and the old File Entity
module.

We can also drop a patch for Media since the change has been merged
in the newer version of the module.