2256: Revert geting ratings from Openlist
Getting ratings from OpenList will increase the number of requests
for the on the service. We do not have any proper form of caching
for the service and worry it will not be able to handled the
increase.

Revert "Merge pull request #694 from b14/b2256"

This reverts commit efd3e011309d360bab21d1667d879a6a5702f07c, reversing
changes made to 29874b6141583d8624eb054a8bc23caffc59b65f.