2819: Hardcode DIBS capture URL
The URL is part of the API, changing it requires insight into the API,
and whether the new URL supports the same API. And DIBS is not likely
to change it without a very good reason (which will most likely
require developer intervention anyway).

And forgetting this setting can cost a lot of money, so let's not put
that kind of pressure on the end user.