Refactor Field/filters and BooleanStatementGroup
Keep the joining operators to the group to simplify adding fields. Having the
operator on the field was confusing as it only applied to the field coming
before the field in question which made the code hard to read.

Instead we now require the developer to use a group if control over the boolean
operator is needed, in all other cases we AND things together

BBS-126