3147: Connie reservation to no set default branch
Removed code in the connie_reseration_options_submit
function because neither Alma nor FBS does anything similar,
so Connie shouldn't either. It was meant to save a default branch
if the user haven't got one.

Also added the connie_reservation_default_options function
because it needs to return an empty array. It was missing.