3029: Implement support for multiple ting relation types
We add a field setting which allows for customisation of which relation
type to use for the reference. The module defined a relation type for
itself but this way we can use the module for other relations as well.

We set the default relation type to ting_reference and add arguments
to functions with ting_reference as default value to preserve backwards
compatibility.