1404: Place p2b field correctly in field group
The field was misplaced in the field_group export from ding_event.
Since ding_event shouldn't really know anything about p2b, a better
approach is to add this in field_group_info_alter.
Also, the field was shown in the wrong view mode. Following the
placement of the other ticket button, it's now shown in default view
mode only.