3147: Renamed context files
The context files were inadequately named from first behat
project. Renamed to libContext for library and often used functions,
and basic behaviour like logging in etc.
P2 renamed to StepsContext - here we will put the steps which are
not generally reusable.
The older naming was more relevant for P2 development. Now there will
no longer be confusion as to whether this is the new regression suite
or the old P2 suite.