3029: Reset covers for entities when custom covers are altered
We react to the relation being created/updated/deleted - not the
entity. When the custom cover is deleted then the relation is deleted
beforehand. This means it is not available for us to query when
deleting the cover and thus we are not able to determine the id
of the Ting object being referenced.