3147: Added accept cookie and ask librarian
The cookie acceptance and ask a librarian - if enabled - can block
parts of the screen where we want to work. So with
Given I accept cookies
we can click both away.
This also introduces css-locator variable (cssStr-array), which at
some point will have to be read in from a file or similar.
Also log_msg and log_timestamp are introduced, so we can add things
to the logfile in an easy way + a shortcut for getPage().