3147: Added search test for openscan
When 3 letters are typed into the search field openscan should provide
some suggestions. This is configurable, and this check expects it to be
enabled.  It cannot really test anything other than it can find
something, and it just lists what it finds in the log file, for the
tester to confirm against configuration.