3147: Added Search on enter test
When typing into the search field and pressing enter, a search should
begin. This check @seek005 is doing just that and ends with scooping
the result. That will fail if no search happened. It is a bit fragile
as it is hardcoded on "Harry", but I assume all libraries will have
at least one book with that criteria.