3147: camelCase and english error msg
After review changed function names to camelCase and also translated
all error messages to english from a mix of danish and english.
During testing found a problem that searching for Å gives hits of
double-A
like AA or Aa. So expanded the search to "Åer", to ensure hits
because it is very common and never spelled with double-a's.