Move Opensearch configuration form to Opensearch module
Install the Opensearch module by default. This makes the most sense
for our majority of Danish libraries. This also means that we can
remove the install task. Provider settings are added as install tasks
automatically.

We leave a placeholder in the original ting settings form in case users
get confused.

Note that the actual configuration variables have not been moved yet.

BBS-146