Remove hard dependency on ting_field_search
Instead of having ding_ting_frontend depend on ting_field_search we
enable the module during the installation. This will allow installations
which do not want to use ting_field_search to disable the module.

Existing installations should not be affected by this as the module is
already enabled. New installations will have the module enabled by
default.

DDBasic references ting_field_search variables but only as checks
For whether variables are set. Since the default is that this is not the
case then having the module disabled will not matter.

BBS-150