Move search profile definitions to ting_field_search
ding_ting_frontend no longer depends on ting_field_search. Consequently
we move the search profile exports to ting_field_search. This ensures 
that modules do not contain irrelevant exports. This could cause
problems in future Feature exports.

Having the definitions in the module which also exposes the exportable
element is not a problem. Features will still be able to work with the 
exports. CTools exportable can be disabled so sites using this module
can disable the existing profiles and provide their own.

BBS-150