Implement search sources retrieval as a provider call
This allows providers to supply a list of known sources for objects
and collections.

This PR replaces the current implementation of source retrieval with a
Provider call search_sources and moves the implementation to the 
opensearch module.

We also provide an implementation of the provider call in the
connie_search module and a test.

BBS-161