3147: Added Makefiles for CirclCI
Copied makefiles from p2-tests-merged and tweaked a bit to fit with the current setup. Whether this will work is to be seen when I push this to the server. I am not sure of any other way of testing it.
The Makefiles worked before.
Added export of BEHAT_PARAMS and setting of it - possibly not with the correct URL - and path names of makefiles. Everything else should work.