3147: Added php to scrutinizer.yml
Because the behat suite has all its code in .php files the php extension is now added to the .scrutinizer.yml file. It may later have to be taken out, as there may be other files with .php that has this far been living 'under the radar'. However, those will be ignored - only files under ding2/tests/behat are to be scrutinized and fixed.