3147: Fixed search result sorting checking
There was a bug in the check-procedures in that it didn't
return anything if the sorting actually passed the checks.
Also fixed that the result scraping starts by navigating to the page it wants
rather than after scraping each page, so it both uses
the general procedure as well as if a problem occurs, then the
screendump will show the page where the problem
was, not the next page.