3147: Modified circle.yml to build and run behat
By copying and modifying the setup from the original behat installation, the
circle.yml and Makefiles in both Ding2 root and in Ding2/tests/behat should now
be set up to build and run the behat tests after each commit and push.
It will also copy the report as well as the screendumps to a circle-directory
for later inspection.