3147: Report format and path fixes
The format for the report renderer was Twig, which gives awesome reports in html, but
it may be better to run with Behat2 format, as that gives more information about
the run. This is changed in the behat.yml-file.
In circle.yml the path for the reports is now fixed. It attempted to hit the full
path, now it is using relative paths.
In Makefile the use of CIRCLE_INDEX_NODE is now removed, as the current suite does
everything the same on all nodes, and in fact could run on just one.