3147: reports continued
Trying to amend circle.yml and Makefile to run. Seems the report files cannot be located.
Also turns out, that if any script fails, then behat returns 1, which is listed as a fail
in build. Trying now to not save the output, but to let it be listed on the output, simple
as that. Not sure how to deal with screenshots for that.