3147: Disable-setting for screendumps
Until it is figured out how to include screendumps on CircleCI, or if
that awaits Jenkins, there is a disable setting in behat.yml, under LibContext.
This is set to true, which means that no screendumps will be taken.
Removing it, or setting it to false will mean screendumps are stored. That is for
running the suite manually. It would be better to have this set through BEHAT_PARAMS
but it doesn't work and I'm running out of time to figure out why.