3147: Added procedure for creating datafiles
Datafiles / stores for test data can be created
automatically using the step definition Given I
create files xxx from opensearch on relation yyy.
This will create two files xxx_Positive.dat and
xxx_Negative.dat. There can be many in each, or few.
The method will attempt to find only books.
The relation will be f.ex. dbcaddi:hasCreatorDescription,
which of course is not a user friendly value, as
it refers to the opensearch fields. However, this
method is a kind of a 'back office' tool, that can
help with creating and re-creating data stores
for relations after f.ex. updates to the service
content.