3147: added tag @cci for circleCI tests
Apparently not all tests can run on circleCI.
The reasons may be inability to configure the
locally built site correctly, or inability to
get the correct data out.
The tests that now have the tag will be the ones
that should run stable and correctly, provided
the site is still in good shape.