3147: Fixed the reservable algorithm
A reservable material is almost always a book.
The invariant that tried to sought this out was
wrong, and let non-books through. The procedure
for finding reservables is now refactored so it
only reads in books if we only want to find
reservables. Notice, it looks for the first three
letters only: Bog. So Bog med stor skrift is passed
through. These are also reservable.