3147: Fixed login deprecated user use
The use of DrupalContext->User is deprecated and got
marked as failed by CircleCI. To enable login
scenarios a variable to hold the user info was
introduced instead, and the code amended. This included
also a check to see if the site is loaded into
the browser - apparantly the drupalcontext->user setting
did that on its own. So now it opens the base-url
in the browser, if the current content cannot be
identified as a drupal generated page. (Drupal will
be in the meta-tags as clear text).