3147: Testing downgrading selenium to 3.5.3
3 tests (pagination and facets) fails due to strange errors.
Locally they work with selenium 3.8.1, but everything is actually
developed using selenium 3.5.3, so trying to let circleCI use that
and see if there is a difference.
It's odd, these have not failed before. And this is the only change
that could interfere with those tests. They also don't have much in common.
So selenium version is best guess.
Notice until this change these tests actually ran successfully a number of times.