3147: Need marionette to run selenium 3.5.3 on circleci
There seems to be a bug that makes selenium 3.5.3 prefer firefox to
chrome, even though it's stated that chrome is to be used.
By adding to capabilities marionette: true, in behat.yml, then this should
work.
Testing it in this commit.
Created a new profile, circleci, for behat, because locally it just doesn't
want to accept that setting. Tested on circleci-server via ssh though, and
it did seem to want to cooperate better, except that I couldn't access the
site, as it was taken down. New test..