3147: Changed waiting to 5 sec for reservation
For upgrade-fbs it is necessary to wait up to 5
secs to check if the reservation button is being
shown or not. Given it's machine time, this is
added. There is no other thing we can wait for,
because first the element is added to the html as
non-reservable, then it may be changed to reservable. Before it just waited for the element to be added. Now with the 5 sec pause, it is
much more likely we end up with the actual result.