3159: Loading opening hours from multiple subqueues
This way we still support the current use case of loading from a single 
subqueue but allow for multiple if they exist.

It also provides a more solid and elegant approach to handling an empty 
set of subqueues. We avoid typejuggling by comparing the subqueues
string to an int and we test the actual output of our subqueue load
instead of a preemptive variable.