3147: Connie.reservation.inc fixes
The reservation options were blanked at an early stage
of the development of the test suite, because we couldn't
make reservations with it.
The original code is put back in, due to comments that it
is important.
Also set phpdoc for new function to update connie to
handle connie_reservation_default_options for selection
of default branch for the user.