3147: Finished testing on cci
Removed the test on material, and put
scenarios depending on known data to
not be part of the cci suite.
Attempted updating to version 5.0 of the
opensearch service for test, but it seems
to not be working. For now these tests
will have to be handheld only, and not run
on every push.