3147: Taking S026 out of cci
All data dependent scenarios cannot run on
cci as long as the data source there is not
determined. Seems very strange, but likely
the initial configuration of the data source
is not entirely correct or done as I think it
is. For CCI it's important that this is stable
so taking these out until this can be fixed at
a later stage.