3147: Scrutinizer fixes
Initiated from Jespers most recent comment:
Now inserting catches for exception types
so only one type is thrown per function.
This is to overcome the problem of the
scrutinizer which doesn't want to approve
of comments for the exceptions if there is more
than one.
Used inspect function of phpstorm on the code
and made some additional fixes it identified
to make the code more robust.