Migrate extended search fields to search abstraction layer
The current implementation relies on building an OpenSearch specific
search string starting with they query and adding additional fields.
This is then added as a part of the query string before redirecting to
the search results.

For this to work with the abstraction layer we would have to serialize
the query object. This is no trivial task.

Instead we add the fields as filters as arguments on the query string.
Then we can convert them to conditions for Drupal Core search and
finally add them add field- or raw filters when building the query
there.

BBS-247