Core: Amend changes to release commits
There is no reason to keep a history of commits on the release branch
hanging around. Their position in relation to the master branch is not
correct anyway.

Consequently we should amend changes to any existing release commit and
thus in practice overriding it. We already rebase the branch on the
master branch. This ensures that each release commit is only a single
commit away from master.