3329: Port ding_messge to use search abstraction layer
Update the code to use proper methods for retrieving values instead of
magic properties.

Building queries can be greatly simplified by using full text queries
and field filters.

The is no need to check the number of collection objects as we have
already established that the total number of collections is greater
than 0.