3329: Clean up search quoting
After refactoring ding_message to use the search abstraction layer the 
_ting_search_quote function is no longer called outside opensearch.

The function in itself is merely a wrapper around the 
TingSearchCqlDoctor class. To clean things up we move the class to the
opensearch module and call the class directly when needed.

Remove all traces of the search provider quote call. It is now unneeded
and we prefer to simplify things before anyone uses them.