3385: Implement covers module using placeholder.com
This module will generate placeholder images for all Ting entities using
the placeholder.com service. This is useful for development and testing
purposed since some functionality such as carousels depend on covers
being available. Few objects in the test services have proper covers.

Placeholder.com is used as it provides needed features like specifying
image dimensions, colors and text. It also seems to be a stable service
in an area where services might get shut down quickly. For example it
has a company backing it - not a private person.

The module is placed under the ting_covers module as it also provides
a reference implementation of a module which provides covers.