3667: Moving the search bar below the menu.
Redesigning the header, in relation to
search page design. (Release 30).
This commit includes moving the DOM element,
but mostly it means tweaking and moving
the Sass styling. The header elements are
placed with fixed positioning, so this commit
changes a lot of lines, but really just
exchanges the settings between the navigation
wrapper and the pane-search-form.
We also need to adjust the hover menu positioning.