3012: Migrate to Circle CI 2.0
This is a migration of the current CircleCI build defintion from
verion 1.0 to 2.0.

In this migration we try to stay as close to the work being carried
out in the 1.0 while using the workflow features of 2.0. This means
that the build will do the following:

1. Build source code for the site using Drush Make and Gulp
2. Install the site
3. From the state (database) of the installed site run unit tests
and Behat tests
4. Produce a package containing the source code needed to run the
site

Work carried out by individual jobs are persisted using the
workspace with the following folder structure:

- project: Contains our project - source code for the Ding2 profile
- drupal: Contains an entire Drupal site with all code needed to run
the site e.g. Drupal, Ding2 profile, contrib modules, libraries and
processed frontend assets.
- db: Contains database dumps
- test: Contains test output (reports, logs, screenshots etc.)
- build: Contains the packaged project

The change requires a number of changes to the current setup:

- Frontend assets are built using Node 6.14 - not 6.10. There is no
6.10 Docker image available from Node or CircleCI.
- Node SASS is rebuild before processing. This is needed for whatever
reason.
- Tests are run using MySQL 5.7.20. There was no specific MySQL
requirement in the 1.0 build but inspection of the built platform
shows that it uses that specific version so we keep that
- Tests are run using PHP 7.0 without specific patch version.
CircleCI does not provide a PHP image with patch version specified.