3841: Add patch to disable customerror redirect
We use the customerror module to show special pages for 403 og 404. If an
anonymous user tries to visit any page without permission, the module will try
to redirect to this page, if the user logs in later. In our setup, this has
some unfortunate effects:

1. Since we use AJAX in our log in process, the redirect is never triggered
when the user logs in, but rather on a subsequent form submission. This can be
very confusing, if the user is redirected to a seemingly unrelated page after,
for example, renewing a loan.
2. The session variable will cause a session cookie to be saved in the user's
browser and following requests will circumvent our Varnish cache.
3. When this error manifests it will, appearently, often be because of an
AJAX-request that recieved 403, adding to the user's confusing (the page is
completely unrelated).

So in this commit we introduce a custom patch to the customerror module,
that completely disables this behavior.