3012: Create releases from tagged builds
Use ghr to upload builds of taggets commits to GitHub. This follows
the approach from
https://circleci.com/blog/publishing-to-github-releases-via-circleci/

We rename the built file to match the name of the tag before
uploading it. This matches the current practice.

If a tag contains the string rc it is deemed a release candidate and
published as a draft release. Otherwise not.