3012: Split Behat tests over containers based on time spent
Test splitting is currently done in a random but predicable process.
However if all slower tests run in one container that container
becomes a bottleneck for the entire build step. Consequently it makes
sense to use test timings as a basis for distributing tests so that
they are of equal run lengths.

liuggio/fastest outputs Behat tests in the format [file]:[line]. This
output is picked up by CircleCI. CircleCI also processes test results
including timings data.

For the test list and the result data to match we patch the JUnit
result formatter. There does not seem to be an easy way to modify the
output of the formatter and there is little chance of a change
getting into Behat anyway.

Produce a patch anf install patches plugin to implement the change.