3875: Build and push database Docker image for installed site
This is orchestrated by CircleCI.

A database Docker images will allow developers to spin up a site without
going through the installation process which can be slow and tedious.

The image is based on MySQL 5.7.23 as that is the closest version to
what we are using for the rest of our CI process.

The image gets the database dump from the workspace copied into a
specific folder which the database image is configured to load SQL
files from on startup.

To provide a stable starting point we set a fixed username and password.

The Dockerfile is not included in the project as we are not providing a
Full Docker setup at this point in time. Instead we generate it during
the build.

Once the image is built it can be pushed to Docker Hub. However this
only occurs if the CircleCI project is configured with the required
username and password. This way we allow forks to utilize our workflow
if they want to but do not force them into anything to avoid breaking
the build.

Tags are used to handle typical use cases:
1. If the codebase has a tag then the image is tagged accordingly.
2. The master branch is latest official development version of the code.
The corresponding database image is tagged master.
3. For all other situations latest is used.

Update documentation accordingly.