3902: Make CirleCI lint PHP files to find syntax errors
Syntax errors are not caught in files that are not a part of our test
setup e.g. provider implementations. Consequently we lint all PHP files
to find such errors.

We use jakub-onderka/php-parallel-lint as that is currently the most
popular package for this. It also supports specifying which file
extensions to scan. This is important for the Drupal 7 file structure.

The tool is run in a job before the site install to find syntax errors
as early as possible.