3902: Remove test classes with syntax errors
DINGTestPageAbstraction.class.inc:51 contains a syntax error. We remove
the file and all other test files for the module as they are obsolete.
They are not being executed as a part of our current continuous
integration workflow, we are unsure whether they work and they are not
in  line with our current test strategy.