3012: Do not depend on tags for CircleCI releases and Docker images
CircleCI has an odd way of handling jobs which are both filtered by tags
and requiring other jobs. In these cases all precursors must also be
filtered by tags for the job to run.
-> https://discuss.circleci.com/t/builds-for-tags-not-triggering/17681/5

This is hard to handle so instead of requiring tags we require the
release branch (which usually contain our tags) and only publish 
releases if a tag for the commit can be determined.