3703: Fix ekurser date sort
Opensearch is using date field for sorting but ding_ekurser was using
acquisitionDate. Before the introduction of SAL this was no problem,
but now Ting search has to verify that the provider supports the current
sort. Since opensearch is not returning acquisitionDate it's ignored both
when performing search and setting default value.
Even though the Opensearch webservice is supporting acquisitionDate (and also
localAcquisitionDate), the simplest and best solution seems to be just make
ding_ekurser use date field for sort.