3901: Remove tracking of page title
It's decided by the business that the page title has no real traking
value in Webtrekk, since the important information is contained within
the URLs that is tracked. This also frees us from the concern of
provider names appearing in page title in the statistics. We could try
to prevent this, but if for example som third party modules is installed
it could add some pages under /user wich has usernames in page titles
and this would be difficult to control.