3905: Refactor opensearch cache insertion/retrieval
To use static caching in the he current implementation we have to pass
the static around between callers when in fact it should be handled
internally.

To improve this we refactor functions for inserting into and retrieval 
from the cache into a single function which can control its static cache
internally.