3603: Use updated jQuery version on admin theme
We have previously updated the site default version of jQuery to
Use via JQuery Update to 3.1. The module has previously been used
to update the site to 1.7.2 and use the same version for the admin
theme.

However the migration path for the updated version of jQuery Update will
set the version to use for the admin theme to Drupal default (1.4) if no
specific version has been set.

This causes JavaScript to break as some of our code relies on updated
versions of jQuery and calls functions which are not defined in 1.4.

To fix this we run an update hook which configures the admin theme to
use the default version set by jQuery Update (3.1).

It is important that this update hook runs after jQuery Updates own
update hooks which will set the version to Drupal default (1.4).