3603: Downgrade jQuery to 1.7 for the administration interface/theme
The latest version of jQuery causes problems for the Module Filter
module: https://www.drupal.org/project/module_filter/issues/2986511

There is no immediate fix for this and the problem which originally
caused the upgrade is only relevant for patron actions performed in the 
DDBasic theme.

Consequently we downgrade the version of jQuery used to 1.7 (in practice
1.7.2) which is the same version currently used.

We need to ensure that the correct version is set on current and new
sites so we implement both hook_update and update hook_install to
make this change.