3603: Prevent access to changing jQuery version
The version of jQuery that we use can have significant impact on the
functionality of the site and break other JavaScript elements.
Consequently normal users should not be able to change this.

To address this we patch jQuery Update to add a new permission which is
used to determine access to functionality provided by jQuery Update.

The new permission is intentionally not assigned to any roles. This
prevents access to everyone except user 1.