3696: Update Honeypot
Our current version if three years old, so I thought i migh as well
update it as part of this PR. It doesn't change the way the spam
protecion is implemented, but has some performance improvements. It does
also provide a new experimental features with Javascript hanlding that
works with out disabling page cache. As this is experimental and we
already handle the cache pages correctly, I think we should keep using
the old strategy.