4039: Fix search spinner on ting object views
Need to show search spinner when clicking on search links, so make the
selector more specific to target those. This is also fixes the unwanted
search spinner on ding entity buttons like the reserve button.
Remove the ting-collection selector since the class is not used and
there's no need for it when we already target the objects in the
collection.
Use the ting_object class instead of the wrapper, since this is always
present on both collection on object view.