3667: Enable dropdown function, for header search.
In relation to a previous revision where the header
structure was moved around, the functionality of
being able to fold/unfold the search line disappeared.

This commit looks to fix that, and also tries to clean up
all the mess in regards to the various combinations that
the header can be displayed in.