3936: Fix problematic ding_entity_call()
When loan and reservation objects has been in the session cache the
ding_entity_id is changed to the value of DingEntityBase::NULL constant
when the object is unserialized and __wakeup() is called.
This means it will try to make a request with ding_entity_load() with
the value of this constant when getEntity() is called, since we only
check for empty ding_entity_id but DingEntityBase::NULL is not empty.
This seems to have been a serious issue because opensearch takes ALOT of
time to hanlde this request making the status pages really slow.