4214: Add non-entity collection retrieval to search abstraction layer
The entity model can be slow to load and it is not required in all
circumstances.

This update adds a new method for retrieving "raw" collections from
a search result. Callers can use this to avoid collection loading.

Collections are only loaded as entities when specifically requested
by calling the existing method.