4214: Remove obsolete reply_only option for OpenSearch
This option was introduced to allow bypassing of the entity loading
process. Since the conversion is now on demand and can be bypassed
by using the appropriate method there is no reason to keep the option
around.

Reimplement the only place where it was used in Core using the
search abstraction layer.