4214: Allow pluggable rendering of search result snippets
The current implementation uses the teaser rendering of each
corresponding collection entity in the search result.

To allow other modules to customize and potentially avoid the
entity model we introduce a new theme function for rendering each
item and pass the “raw” collection to this function.

The default implementation copies the current behavior by loading
the corresponding collection entity and rendering the teaser view
for this.