4214: Initial implementation of fast search module
This module provides an alternative implementation of the search
result rendering. Instead of using the teaser view of the
corresponding collection entity we instead hardcode the extraction,
ordering and output of each collection.

This way we avoid using the entity model altogether.

By not using the entity model we reduce data loading from Drupal
as well as potentially extra calls to the search engine which would
otherwise need to be avoiding using increasingly complex caching
mechanisms.