4156: Handle ticket URL field synchronization during migration
In ding_place2book_update_7004() all p2b events got their ticket order
URL inserted in the ding_event ticket url field to better integrate
p2b events with the rest of the system. If this update wasn't run with
the old version of the module the p2b URL hasn't been updated, since
the update function was using a function only present in the old
module. We could have rewritten the function to work with in the new
module, but it's easier to just handle it here during migration. So
here we ensure that the ticket URL field is in fact synchronized with
the p2b sales location.