4194: Ensure local ids are addes as string values to request
Datawell ids needs to be strings, since for example the id '06606059'
is not the same as '6606059'. But PHP might have converted some of the
local ids to integers when using them as array keys. For example the id
'20174870' would be converted to integer when using as key. It's
important that the request object is exactly the same when using it as
a base for cache ID, so here we ensure that all local ids are added to
the request as string values.