1679: Change color of comments in Behat formatter
Behat prints the location (file line/class method) of scenarios and
steps, which is nice for debugging. For some reason it also decides to
print comment with black text color, which renders them invisible (but
copy-paste-able) in a terminal with black background (though Circle
uses a black background, it also seem to ignore the black text color).

So for the sanity of those using terminals with The One True
background color, reset the text color of comments to make them visible.